Move TestCopyRecursive into the same package as CopyRecursive
The CopyRecursive function got moved a few months ago, but it's test
didn't. Cleaning this up now before I add more tests in an upcoming PR.
