+++ templates/commands/render/action_append_test.go
@@ -178,9 +178,9 @@ func TestActionAppend(t *testing.T) {
 				},
 			}
 			sp := &stepParams{
-				fs: &errorFS{
+				fs: &common.ErrorFS{
 					FS:          &common.RealFS{},
-					readFileErr: tc.readFileErr,
+					ReadFileErr: tc.readFileErr,
 				},
 				scratchDir: scratchDir,
 				scope:      common.NewScope(tc.inputs),
