+++ templates/commands/render/action_test.go
@@ -19,20 +19,14 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"io/fs"
 	"path/filepath"
-	"strings"
 	"testing"
-	"time"
 
 	"github.com/abcxyz/abc/templates/common"
 	"github.com/abcxyz/abc/templates/model"
 	"github.com/abcxyz/pkg/logging"
 	"github.com/abcxyz/pkg/testutil"
-	"github.com/benbjohnson/clock"
 	"github.com/google/go-cmp/cmp"
-	"github.com/google/go-cmp/cmp/cmpopts"
-	"golang.org/x/exp/slices"
 )
 
 func TestWalkAndModify(t *testing.T) {
@@ -223,10 +217,10 @@ func TestWalkAndModify(t *testing.T) {
 			common.WriteAllDefaultMode(t, scratchDir, tc.initialContents)
 
 			sp := &stepParams{
-				fs: &errorFS{
+				fs: &common.ErrorFS{
 					FS:           &common.RealFS{},
-					readFileErr:  tc.readFileErr,
-					writeFileErr: tc.writeFileErr,
+					ReadFileErr:  tc.readFileErr,
+					WriteFileErr: tc.writeFileErr,
 				},
 				scratchDir: scratchDir,
 				scope:      common.NewScope(nil),
@@ -252,373 +246,7 @@ func TestWalkAndModify(t *testing.T) {
 	}
 }
 
-func TestCopyRecursive(t *testing.T) {
-	t.Parallel()
-
-	cases := []struct {
-		name                  string
-		srcDirContents        map[string]common.ModeAndContents
-		suffix                string
-		dryRun                bool
-		visitor               common.CopyVisitor
-		want                  map[string]common.ModeAndContents
-		wantBackups           map[string]common.ModeAndContents
-		dstDirInitialContents map[string]common.ModeAndContents // only used in the tests for overwriting and backing up
-		mkdirAllErr           error
-		openErr               error
-		openFileErr           error
-		readFileErr           error
-		statErr               error
-		writeFileErr          error
-		wantErr               string
-	}{
-		{
-			name: "simple_success",
-			srcDirContents: map[string]common.ModeAndContents{
-				"file1.txt":      {Mode: 0o600, Contents: "file1 contents"},
-				"dir1/file2.txt": {Mode: 0o600, Contents: "file2 contents"},
-			},
-			want: map[string]common.ModeAndContents{
-				"file1.txt":      {Mode: 0o600, Contents: "file1 contents"},
-				"dir1/file2.txt": {Mode: 0o600, Contents: "file2 contents"},
-			},
-		},
-		{
-			name: "dry_run_should_not_change_anything",
-			srcDirContents: map[string]common.ModeAndContents{
-				"file1.txt":      {Mode: 0o600, Contents: "file1 contents"},
-				"dir1/file2.txt": {Mode: 0o600, Contents: "file2 contents"},
-			},
-			dryRun:      true,
-			openFileErr: fmt.Errorf("OpenFile shouldn't be called in dry run mode"),
-			mkdirAllErr: fmt.Errorf("MkdirAll shouldn't be called in dry run mode"),
-		},
-		{
-			name: "dry_run_with_overwrite_doesnt_make_backups",
-			srcDirContents: map[string]common.ModeAndContents{
-				"file1.txt": {Mode: 0o600, Contents: "new contents"},
-			},
-			dstDirInitialContents: map[string]common.ModeAndContents{
-				"file1.txt": {Mode: 0o600, Contents: "old contents"},
-			},
-			visitor: func(relPath string, de fs.DirEntry) (common.CopyHint, error) {
-				return common.CopyHint{
-					BackupIfExists: true,
-					Overwrite:      true,
-				}, nil
-			},
-			want: map[string]common.ModeAndContents{
-				"file1.txt": {Mode: 0o600, Contents: "old contents"},
-			},
-
-			dryRun: true,
-		},
-		{
-			name: "dry_run_without_overwrite_should_detect_conflicting_files",
-			dstDirInitialContents: map[string]common.ModeAndContents{
-				"file1.txt": {Mode: 0o600, Contents: "old contents"},
-			},
-			srcDirContents: map[string]common.ModeAndContents{
-				"file1.txt":      {Mode: 0o600, Contents: "new contents"},
-				"dir1/file2.txt": {Mode: 0o600, Contents: "file2 contents"},
-			},
-			want: map[string]common.ModeAndContents{
-				"file1.txt": {Mode: 0o600, Contents: "old contents"},
-			},
-			dryRun: true,
-			visitor: func(relPath string, de fs.DirEntry) (common.CopyHint, error) {
-				return common.CopyHint{
-					Overwrite: false,
-				}, nil
-			},
-			openFileErr: fmt.Errorf("OpenFile shouldn't be called in dry run mode"),
-			mkdirAllErr: fmt.Errorf("MkdirAll shouldn't be called in dry run mode"),
-			wantErr:     "file file1.txt already exists and overwriting was not enabled",
-		},
-		{
-			name: "owner_execute_bit_should_be_preserved",
-			srcDirContents: map[string]common.ModeAndContents{
-				"myfile1.txt": {Mode: 0o600, Contents: "my file contents"},
-				"myfile2.txt": {Mode: 0o700, Contents: "my file contents"},
-			},
-			want: map[string]common.ModeAndContents{
-				"myfile1.txt": {Mode: 0o600, Contents: "my file contents"},
-				"myfile2.txt": {Mode: 0o700, Contents: "my file contents"},
-			},
-		},
-		{
-			name:   "copying_a_file_rather_than_directory_should_work",
-			suffix: "myfile1.txt",
-			srcDirContents: map[string]common.ModeAndContents{
-				"myfile1.txt": {Mode: 0o600, Contents: "my file contents"},
-			},
-			want: map[string]common.ModeAndContents{
-				"myfile1.txt": {Mode: 0o600, Contents: "my file contents"},
-			},
-		},
-		{
-			name: "deep_directories_should_work",
-			srcDirContents: map[string]common.ModeAndContents{
-				"dir/dir/dir/dir/dir/file.txt": {Mode: 0o600, Contents: "file contents"},
-			},
-			want: map[string]common.ModeAndContents{
-				"dir/dir/dir/dir/dir/file.txt": {Mode: 0o600, Contents: "file contents"},
-			},
-		},
-		{
-			name: "directories_with_several_files_should_work",
-			srcDirContents: map[string]common.ModeAndContents{
-				"f1.txt": {Mode: 0o600, Contents: "abc"},
-				"f2.txt": {Mode: 0o600, Contents: "def"},
-				"f3.txt": {Mode: 0o600, Contents: "ghi"},
-				"f4.txt": {Mode: 0o600, Contents: "jkl"},
-				"f5.txt": {Mode: 0o600, Contents: "mno"},
-				"f6.txt": {Mode: 0o600, Contents: "pqr"},
-				"f7.txt": {Mode: 0o600, Contents: "stu"},
-				"f8.txt": {Mode: 0o600, Contents: "vwx"},
-				"f9.txt": {Mode: 0o600, Contents: "yz"},
-			},
-			want: map[string]common.ModeAndContents{
-				"f1.txt": {Mode: 0o600, Contents: "abc"},
-				"f2.txt": {Mode: 0o600, Contents: "def"},
-				"f3.txt": {Mode: 0o600, Contents: "ghi"},
-				"f4.txt": {Mode: 0o600, Contents: "jkl"},
-				"f5.txt": {Mode: 0o600, Contents: "mno"},
-				"f6.txt": {Mode: 0o600, Contents: "pqr"},
-				"f7.txt": {Mode: 0o600, Contents: "stu"},
-				"f8.txt": {Mode: 0o600, Contents: "vwx"},
-				"f9.txt": {Mode: 0o600, Contents: "yz"},
-			},
-		},
-		{
-			name: "overwriting_with_overwrite_true_should_succeed",
-			srcDirContents: map[string]common.ModeAndContents{
-				"file1.txt": {Mode: 0o600, Contents: "new contents"},
-			},
-			dstDirInitialContents: map[string]common.ModeAndContents{
-				"file1.txt": {Mode: 0o600, Contents: "old contents"},
-			},
-			visitor: func(relPath string, de fs.DirEntry) (common.CopyHint, error) {
-				return common.CopyHint{
-					Overwrite: true,
-				}, nil
-			},
-			want: map[string]common.ModeAndContents{
-				"file1.txt": {Mode: 0o600, Contents: "new contents"},
-			},
-		},
-		{
-			name: "overwriting_with_overwrite_false_should_fail",
-			srcDirContents: map[string]common.ModeAndContents{
-				"file1.txt": {Mode: 0o600, Contents: "new contents"},
-			},
-			dstDirInitialContents: map[string]common.ModeAndContents{
-				"file1.txt": {Mode: 0o600, Contents: "old contents"},
-			},
-			want: map[string]common.ModeAndContents{
-				"file1.txt": {Mode: 0o600, Contents: "old contents"},
-			},
-			wantErr: "overwriting was not enabled",
-		},
-		{
-			name: "overwriting_dir_with_child_file_should_fail",
-			visitor: func(relPath string, de fs.DirEntry) (common.CopyHint, error) {
-				return common.CopyHint{
-					Overwrite: true,
-				}, nil
-			},
-			srcDirContents: map[string]common.ModeAndContents{
-				"a": {Mode: 0o600, Contents: "file contents"},
-			},
-			dstDirInitialContents: map[string]common.ModeAndContents{
-				"a/b.txt": {Mode: 0o600, Contents: "file contents"},
-			},
-			want: map[string]common.ModeAndContents{
-				"a/b.txt": {Mode: 0o600, Contents: "file contents"},
-			},
-			wantErr: "cannot overwrite a directory with a file of the same name",
-		},
-		{
-			name: "overwriting_file_with_dir_should_fail",
-			visitor: func(relPath string, de fs.DirEntry) (common.CopyHint, error) {
-				return common.CopyHint{
-					Overwrite: true,
-				}, nil
-			},
-			srcDirContents: map[string]common.ModeAndContents{
-				"a/b.txt": {Mode: 0o600, Contents: "file contents"},
-			},
-			dstDirInitialContents: map[string]common.ModeAndContents{
-				"a": {Mode: 0o600, Contents: "file contents"},
-			},
-			want: map[string]common.ModeAndContents{
-				"a": {Mode: 0o600, Contents: "file contents"},
-			},
-			wantErr: "cannot overwrite a file with a directory of the same name",
-		},
-		{
-			name: "skipped_files",
-			srcDirContents: map[string]common.ModeAndContents{
-				"file1.txt":        {Mode: 0o600, Contents: "file1 contents"},
-				"dir1/file2.txt":   {Mode: 0o600, Contents: "file2 contents"},
-				"skip1.txt":        {Mode: 0o600, Contents: "skip1.txt contents"},
-				"subdir/skip2.txt": {Mode: 0o600, Contents: "skip2.txt contents"},
-				"z.txt":            {Mode: 0o600, Contents: "z.txt contents"},
-			},
-			visitor: func(relPath string, de fs.DirEntry) (common.CopyHint, error) {
-				return common.CopyHint{
-					Skip: slices.Contains([]string{"skip1.txt", filepath.Join("subdir", "skip2.txt")}, relPath),
-				}, nil
-			},
-			want: map[string]common.ModeAndContents{
-				"file1.txt":      {Mode: 0o600, Contents: "file1 contents"},
-				"dir1/file2.txt": {Mode: 0o600, Contents: "file2 contents"},
-				"z.txt":          {Mode: 0o600, Contents: "z.txt contents"},
-			},
-		},
-		{
-			name: "skipped_directory_skips_all_subsequent",
-			srcDirContents: map[string]common.ModeAndContents{
-				"file1.txt":          {Mode: 0o600, Contents: "file1 contents"},
-				"subdir/file2.txt":   {Mode: 0o600, Contents: "file2 contents"},
-				"subdir/file3.txt":   {Mode: 0o600, Contents: "file3 contents"},
-				"otherdir/file4.txt": {Mode: 0o600, Contents: "file4 contents"},
-			},
-			visitor: func(relPath string, de fs.DirEntry) (common.CopyHint, error) {
-				if relPath == "subdir" {
-					return common.CopyHint{
-						Skip: true,
-					}, nil
-				}
-				if strings.HasPrefix(relPath, "subdir/") {
-					panic("no children of subdir/ should have been walked")
-				}
-				return common.CopyHint{}, nil
-			},
-			want: map[string]common.ModeAndContents{
-				"file1.txt":          {Mode: 0o600, Contents: "file1 contents"},
-				"otherdir/file4.txt": {Mode: 0o600, Contents: "file4 contents"},
-			},
-		},
-		{
-			name: "backup_existing",
-			srcDirContents: map[string]common.ModeAndContents{
-				"file1.txt":        {Mode: 0o600, Contents: "file1 new contents"},
-				"subdir/file2.txt": {Mode: 0o600, Contents: "file2 new contents"},
-			},
-			dstDirInitialContents: map[string]common.ModeAndContents{
-				"file1.txt":        {Mode: 0o600, Contents: "file1 old contents"},
-				"subdir/file2.txt": {Mode: 0o600, Contents: "file2 old contents"},
-			},
-			visitor: func(relPath string, de fs.DirEntry) (common.CopyHint, error) {
-				return common.CopyHint{
-					BackupIfExists: true,
-					Overwrite:      true,
-				}, nil
-			},
-			want: map[string]common.ModeAndContents{
-				"file1.txt":        {Mode: 0o600, Contents: "file1 new contents"},
-				"subdir/file2.txt": {Mode: 0o600, Contents: "file2 new contents"},
-			},
-			wantBackups: map[string]common.ModeAndContents{
-				"file1.txt":        {Mode: 0o600, Contents: "file1 old contents"},
-				"subdir/file2.txt": {Mode: 0o600, Contents: "file2 old contents"},
-			},
-		},
-		{
-			name: "MkdirAll error should be returned",
-			srcDirContents: map[string]common.ModeAndContents{
-				"dir/file.txt": {Mode: 0o600, Contents: "file1 contents"},
-			},
-			mkdirAllErr: fmt.Errorf("fake error"),
-			wantErr:     "MkdirAll(): fake error",
-		},
-		{
-			name: "Open error should be returned",
-			srcDirContents: map[string]common.ModeAndContents{
-				"dir/file.txt": {Mode: 0o600, Contents: "file1 contents"},
-			},
-			openErr: fmt.Errorf("fake error"),
-			wantErr: "fake error", // This error comes from WalkDir, not from our own code, so it doesn't have an "Open():" at the beginning
-		},
-		{
-			name: "OpenFile error should be returned",
-			srcDirContents: map[string]common.ModeAndContents{
-				"dir/file.txt": {Mode: 0o600, Contents: "file1 contents"},
-			},
-			openFileErr: fmt.Errorf("fake error"),
-			wantErr:     "OpenFile(): fake error",
-		},
-		{
-			name: "Stat error should be returned",
-			srcDirContents: map[string]common.ModeAndContents{
-				"dir/file.txt": {Mode: 0o600, Contents: "file1 contents"},
-			},
-			statErr: fmt.Errorf("fake error"),
-			wantErr: "fake error", // This error comes from WalkDir, not from our own code, so it doesn't have a "Stat():" at the beginning
-		},
-	}
-
-	for _, tc := range cases {
-		tc := tc
-
-		t.Run(tc.name, func(t *testing.T) {
-			t.Parallel()
-
-			tempDir := t.TempDir()
-			fromDir := filepath.Join(tempDir, "from_dir")
-			toDir := filepath.Join(tempDir, "to_dir")
-			backupDir := filepath.Join(tempDir, "backups")
-
-			common.WriteAll(t, fromDir, tc.srcDirContents)
-
-			from := fromDir
-			to := toDir
-			if tc.suffix != "" {
-				from = filepath.Join(fromDir, tc.suffix)
-				to = filepath.Join(toDir, tc.suffix)
-			}
-			common.WriteAll(t, toDir, tc.dstDirInitialContents)
-			fs := &errorFS{
-				FS: &common.RealFS{},
-
-				mkdirAllErr: tc.mkdirAllErr,
-				openErr:     tc.openErr,
-				openFileErr: tc.openFileErr,
-				statErr:     tc.statErr,
-			}
-			ctx := context.Background()
-
-			clk := clock.NewMock()
-			const unixTime = 1688609125
-			clk.Set(time.Unix(unixTime, 0)) // Arbitrary timestamp
-
-			err := common.CopyRecursive(ctx, &model.ConfigPos{}, &common.CopyParams{
-				BackupDirMaker: func(rf common.FS) (string, error) { return backupDir, nil },
-				SrcRoot:        from,
-				DstRoot:        to,
-				DryRun:         tc.dryRun,
-				RFS:            fs,
-				Visitor:        tc.visitor,
-			})
-			if diff := testutil.DiffErrString(err, tc.wantErr); diff != "" {
-				t.Error(diff)
-			}
-
-			got := common.LoadDirContents(t, toDir)
-			if diff := cmp.Diff(got, tc.want, common.CmpFileMode, cmpopts.EquateEmpty()); diff != "" {
-				t.Errorf("destination directory was not as expected (-got,+want): %s", diff)
-			}
-
-			gotBackups := common.LoadDirContents(t, backupDir)
-			if diff := cmp.Diff(gotBackups, tc.wantBackups, common.CmpFileMode, cmpopts.EquateEmpty()); diff != "" {
-				t.Errorf("backups directory was not as expected (-got,+want): %s", diff)
-			}
-		})
-	}
-}
-
-func TestParseAndExecute(t *testing.T) {
+func TestParseAndExecuteGoTmpl(t *testing.T) {
 	t.Parallel()
 
 	cases := []struct {
